refactor: extract routes JSON deserialization helpers into import-safe module
Move parse_json_fields, FINDING_JSON_FIELDS, deserialize_finding_rows and
deserialize_asset_service_rows out of routes.py into a new small module
routes_json_helpers.py so unit tests can import and exercise the real
production functions without pulling in the heavy routes.py import chain
(FastAPI, reporting, xhtml2pdf).

routes.py re-exports the helpers so all existing call sites and downstream
imports keep working unchanged.

The dedicated unit test file now imports from routes_json_helpers directly,
so a regression in the actual implementation is caught by these tests instead
of testing frozen copies that could silently drift.
